+++ lib/src/codepages/unicode_iso8859.g.dart
@@ -1,0 +1,283 @@
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// GENERATED FILE, DO NOT EDIT.
+
+// Tables for ISO-8859 code page encoding of Unicode text,
+// using the [authoritative](https://unicode.org/Public/MAPPINGS/ISO8859/ReadMe.txt)
+// Unicode Consortium character
+// [mappings](https://unicode.org/Public/MAPPINGS/ISO8859/ "Mapping repository")
+
+// ignore_for_file: missing_whitespace_between_adjacent_strings
+
+// Generated content contains constant strings named `iso8859_1` through
+// `iso8859_16` (no `_12`), each containing 256 Unicode characters
+// from the basic multilingual plane (code point below U+10000).
+
+// -- BEGIN GENERATED CONTENT --
+// Generated (2023-11-07T14:27) by tools/update_unicode_iso8859_tables.dart
+
+// Characters of ISO-8859-2.
+const iso8859_2 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-2.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u0104\u02D8\u0141\xA4\u013D\u015A\xA7'
+    '\xA8\u0160\u015E\u0164\u0179\xAD\u017D\u017B'
+    '\xB0\u0105\u02DB\u0142\xB4\u013E\u015B\u02C7'
+    '\xB8\u0161\u015F\u0165\u017A\u02DD\u017E\u017C'
+    '\u0154\xC1\xC2\u0102\xC4\u0139\u0106\xC7'
+    '\u010C\xC9\u0118\xCB\u011A\xCD\xCE\u010E'
+    '\u0110\u0143\u0147\xD3\xD4\u0150\xD6\xD7'
+    '\u0158\u016E\xDA\u0170\xDC\xDD\u0162\xDF'
+    '\u0155\xE1\xE2\u0103\xE4\u013A\u0107\xE7'
+    '\u010D\xE9\u0119\xEB\u011B\xED\xEE\u010F'
+    '\u0111\u0144\u0148\xF3\xF4\u0151\xF6\xF7'
+    '\u0159\u016F\xFA\u0171\xFC\xFD\u0163\u02D9';
+
+// Characters of ISO-8859-3.
+const iso8859_3 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-3.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u0126\u02D8\xA3\xA4\uFFFD\u0124\xA7'
+    '\xA8\u0130\u015E\u011E\u0134\xAD\uFFFD\u017B'
+    '\xB0\u0127\xB2\xB3\xB4\xB5\u0125\xB7'
+    '\xB8\u0131\u015F\u011F\u0135\xBD\uFFFD\u017C'
+    '\xC0\xC1\xC2\uFFFD\xC4\u010A\u0108\xC7\xC8\xC9\xCA\xCB\xCC\xCD\xCE\xCF'
+    '\uFFFD\xD1\xD2\xD3\xD4\u0120\xD6\xD7\u011C\xD9\xDA\xDB\xDC\u016C\u015C\xDF'
+    '\xE0\xE1\xE2\uFFFD\xE4\u010B\u0109\xE7\xE8\xE9\xEA\xEB\xEC\xED\xEE\xEF'
+    '\uFFFD\xF1\xF2\xF3\xF4\u0121\xF6\xF7'
+    '\u011D\xF9\xFA\xFB\xFC\u016D\u015D\u02D9';
+
+// Characters of ISO-8859-4.
+const iso8859_4 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-4.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u0104\u0138\u0156\xA4\u0128\u013B\xA7'
+    '\xA8\u0160\u0112\u0122\u0166\xAD\u017D\xAF'
+    '\xB0\u0105\u02DB\u0157\xB4\u0129\u013C\u02C7'
+    '\xB8\u0161\u0113\u0123\u0167\u014A\u017E\u014B'
+    '\u0100\xC1\xC2\xC3\xC4\xC5\xC6\u012E'
+    '\u010C\xC9\u0118\xCB\u0116\xCD\xCE\u012A'
+    '\u0110\u0145\u014C\u0136\xD4\xD5\xD6\xD7'
+    '\xD8\u0172\xDA\xDB\xDC\u0168\u016A\xDF\u0101\xE1\xE2\xE3\xE4\xE5\xE6\u012F'
+    '\u010D\xE9\u0119\xEB\u0117\xED\xEE\u012B'
+    '\u0111\u0146\u014D\u0137\xF4\xF5\xF6\xF7'
+    '\xF8\u0173\xFA\xFB\xFC\u0169\u016B\u02D9';
+
+// Characters of ISO-8859-5.
+const iso8859_5 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-5.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u0401\u0402\u0403\u0404\u0405\u0406\u0407'
+    '\u0408\u0409\u040A\u040B\u040C\xAD\u040E\u040F'
+    '\u0410\u0411\u0412\u0413\u0414\u0415\u0416\u0417'
+    '\u0418\u0419\u041A\u041B\u041C\u041D\u041E\u041F'
+    '\u0420\u0421\u0422\u0423\u0424\u0425\u0426\u0427'
+    '\u0428\u0429\u042A\u042B\u042C\u042D\u042E\u042F'
+    '\u0430\u0431\u0432\u0433\u0434\u0435\u0436\u0437'
+    '\u0438\u0439\u043A\u043B\u043C\u043D\u043E\u043F'
+    '\u0440\u0441\u0442\u0443\u0444\u0445\u0446\u0447'
+    '\u0448\u0449\u044A\u044B\u044C\u044D\u044E\u044F'
+    '\u2116\u0451\u0452\u0453\u0454\u0455\u0456\u0457'
+    '\u0458\u0459\u045A\u045B\u045C\xA7\u045E\u045F';
+
+// Characters of ISO-8859-6.
+const iso8859_6 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-6.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\uFFFD\uFFFD\uFFFD\xA4\uFFFD\uFFFD\uFFFD'
+    '\uFFFD\uFFFD\uFFFD\uFFFD\u060C\xAD\uFFFD\uFFFD'
+    '\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD'
+    '\uFFFD\uFFFD\uFFFD\u061B\uFFFD\uFFFD\uFFFD\u061F'
+    '\uFFFD\u0621\u0622\u0623\u0624\u0625\u0626\u0627'
+    '\u0628\u0629\u062A\u062B\u062C\u062D\u062E\u062F'
+    '\u0630\u0631\u0632\u0633\u0634\u0635\u0636\u0637'
+    '\u0638\u0639\u063A\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD'
+    '\u0640\u0641\u0642\u0643\u0644\u0645\u0646\u0647'
+    '\u0648\u0649\u064A\u064B\u064C\u064D\u064E\u064F'
+    '\u0650\u0651\u0652\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD'
+    '\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD';
+
+// Characters of ISO-8859-7.
+const iso8859_7 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-7.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u2018\u2019\xA3\u20AC\u20AF\xA6\xA7'
+    '\xA8\xA9\u037A\xAB\xAC\xAD\uFFFD\u2015'
+    '\xB0\xB1\xB2\xB3\u0384\u0385\u0386\xB7'
+    '\u0388\u0389\u038A\xBB\u038C\xBD\u038E\u038F'
+    '\u0390\u0391\u0392\u0393\u0394\u0395\u0396\u0397'
+    '\u0398\u0399\u039A\u039B\u039C\u039D\u039E\u039F'
+    '\u03A0\u03A1\uFFFD\u03A3\u03A4\u03A5\u03A6\u03A7'
+    '\u03A8\u03A9\u03AA\u03AB\u03AC\u03AD\u03AE\u03AF'
+    '\u03B0\u03B1\u03B2\u03B3\u03B4\u03B5\u03B6\u03B7'
+    '\u03B8\u03B9\u03BA\u03BB\u03BC\u03BD\u03BE\u03BF'
+    '\u03C0\u03C1\u03C2\u03C3\u03C4\u03C5\u03C6\u03C7'
+    '\u03C8\u03C9\u03CA\u03CB\u03CC\u03CD\u03CE\uFFFD';
+
+// Characters of ISO-8859-8.
+const iso8859_8 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-8.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\uFFFD\xA2\xA3\xA4\xA5\xA6\xA7\xA8\xA9\xD7\xAB\xAC\xAD\xAE\xAF'
+    '\xB0\xB1\xB2\xB3\xB4\xB5\xB6\xB7\xB8\xB9\xF7\xBB\xBC\xBD\xBE\uFFFD'
+    '\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD'
+    '\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD'
+    '\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD'
+    '\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\uFFFD\u2017'
+    '\u05D0\u05D1\u05D2\u05D3\u05D4\u05D5\u05D6\u05D7'
+    '\u05D8\u05D9\u05DA\u05DB\u05DC\u05DD\u05DE\u05DF'
+    '\u05E0\u05E1\u05E2\u05E3\u05E4\u05E5\u05E6\u05E7'
+    '\u05E8\u05E9\u05EA\uFFFD\uFFFD\u200E\u200F\uFFFD';
+
+// Characters of ISO-8859-9.
+const iso8859_9 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-9.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\xA1\xA2\xA3\xA4\xA5\xA6\xA7\xA8\xA9\xAA\xAB\xAC\xAD\xAE\xAF'
+    '\xB0\xB1\xB2\xB3\xB4\xB5\xB6\xB7\xB8\xB9\xBA\xBB\xBC\xBD\xBE\xBF'
+    '\xC0\xC1\xC2\xC3\xC4\xC5\xC6\xC7\xC8\xC9\xCA\xCB\xCC\xCD\xCE\xCF'
+    '\u011E\xD1\xD2\xD3\xD4\xD5\xD6\xD7\xD8\xD9\xDA\xDB\xDC\u0130\u015E\xDF'
+    '\xE0\xE1\xE2\xE3\xE4\xE5\xE6\xE7\xE8\xE9\xEA\xEB\xEC\xED\xEE\xEF'
+    '\u011F\xF1\xF2\xF3\xF4\xF5\xF6\xF7\xF8\xF9\xFA\xFB\xFC\u0131\u015F\xFF';
+
+// Characters of ISO-8859-10.
+const iso8859_10 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-10.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u0104\u0112\u0122\u012A\u0128\u0136\xA7'
+    '\u013B\u0110\u0160\u0166\u017D\xAD\u016A\u014A'
+    '\xB0\u0105\u0113\u0123\u012B\u0129\u0137\xB7'
+    '\u013C\u0111\u0161\u0167\u017E\u2015\u016B\u014B'
+    '\u0100\xC1\xC2\xC3\xC4\xC5\xC6\u012E\u010C\xC9\u0118\xCB\u0116\xCD\xCE\xCF'
+    '\xD0\u0145\u014C\xD3\xD4\xD5\xD6\u0168\xD8\u0172\xDA\xDB\xDC\xDD\xDE\xDF'
+    '\u0101\xE1\xE2\xE3\xE4\xE5\xE6\u012F\u010D\xE9\u0119\xEB\u0117\xED\xEE\xEF'
+    '\xF0\u0146\u014D\xF3\xF4\xF5\xF6\u0169\xF8\u0173\xFA\xFB\xFC\xFD\xFE\u0138';
+
+// Characters of ISO-8859-11.
+const iso8859_11 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-11.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u0E01\u0E02\u0E03\u0E04\u0E05\u0E06\u0E07'
+    '\u0E08\u0E09\u0E0A\u0E0B\u0E0C\u0E0D\u0E0E\u0E0F'
+    '\u0E10\u0E11\u0E12\u0E13\u0E14\u0E15\u0E16\u0E17'
+    '\u0E18\u0E19\u0E1A\u0E1B\u0E1C\u0E1D\u0E1E\u0E1F'
+    '\u0E20\u0E21\u0E22\u0E23\u0E24\u0E25\u0E26\u0E27'
+    '\u0E28\u0E29\u0E2A\u0E2B\u0E2C\u0E2D\u0E2E\u0E2F'
+    '\u0E30\u0E31\u0E32\u0E33\u0E34\u0E35\u0E36\u0E37'
+    '\u0E38\u0E39\u0E3A\uFFFD\uFFFD\uFFFD\uFFFD\u0E3F'
+    '\u0E40\u0E41\u0E42\u0E43\u0E44\u0E45\u0E46\u0E47'
+    '\u0E48\u0E49\u0E4A\u0E4B\u0E4C\u0E4D\u0E4E\u0E4F'
+    '\u0E50\u0E51\u0E52\u0E53\u0E54\u0E55\u0E56\u0E57'
+    '\u0E58\u0E59\u0E5A\u0E5B\uFFFD\uFFFD\uFFFD\uFFFD';
+
+// Characters of ISO-8859-13.
+const iso8859_13 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-13.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u201D\xA2\xA3\xA4\u201E\xA6\xA7\xD8\xA9\u0156\xAB\xAC\xAD\xAE\xC6'
+    '\xB0\xB1\xB2\xB3\u201C\xB5\xB6\xB7\xF8\xB9\u0157\xBB\xBC\xBD\xBE\xE6'
+    '\u0104\u012E\u0100\u0106\xC4\xC5\u0118\u0112'
+    '\u010C\xC9\u0179\u0116\u0122\u0136\u012A\u013B'
+    '\u0160\u0143\u0145\xD3\u014C\xD5\xD6\xD7'
+    '\u0172\u0141\u015A\u016A\xDC\u017B\u017D\xDF'
+    '\u0105\u012F\u0101\u0107\xE4\xE5\u0119\u0113'
+    '\u010D\xE9\u017A\u0117\u0123\u0137\u012B\u013C'
+    '\u0161\u0144\u0146\xF3\u014D\xF5\xF6\xF7'
+    '\u0173\u0142\u015B\u016B\xFC\u017C\u017E\u2019';
+
+// Characters of ISO-8859-14.
+const iso8859_14 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-14.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u1E02\u1E03\xA3\u010A\u010B\u1E0A\xA7'
+    '\u1E80\xA9\u1E82\u1E0B\u1EF2\xAD\xAE\u0178'
+    '\u1E1E\u1E1F\u0120\u0121\u1E40\u1E41\xB6\u1E56'
+    '\u1E81\u1E57\u1E83\u1E60\u1EF3\u1E84\u1E85\u1E61'
+    '\xC0\xC1\xC2\xC3\xC4\xC5\xC6\xC7\xC8\xC9\xCA\xCB\xCC\xCD\xCE\xCF'
+    '\u0174\xD1\xD2\xD3\xD4\xD5\xD6\u1E6A\xD8\xD9\xDA\xDB\xDC\xDD\u0176\xDF'
+    '\xE0\xE1\xE2\xE3\xE4\xE5\xE6\xE7\xE8\xE9\xEA\xEB\xEC\xED\xEE\xEF'
+    '\u0175\xF1\xF2\xF3\xF4\xF5\xF6\u1E6B\xF8\xF9\xFA\xFB\xFC\xFD\u0177\xFF';
+
+// Characters of ISO-8859-15.
+const iso8859_15 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-15.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\xA1\xA2\xA3\u20AC\xA5\u0160\xA7\u0161\xA9\xAA\xAB\xAC\xAD\xAE\xAF'
+    '\xB0\xB1\xB2\xB3\u017D\xB5\xB6\xB7\u017E\xB9\xBA\xBB\u0152\u0153\u0178\xBF'
+    '\xC0\xC1\xC2\xC3\xC4\xC5\xC6\xC7\xC8\xC9\xCA\xCB\xCC\xCD\xCE\xCF'
+    '\xD0\xD1\xD2\xD3\xD4\xD5\xD6\xD7\xD8\xD9\xDA\xDB\xDC\xDD\xDE\xDF'
+    '\xE0\xE1\xE2\xE3\xE4\xE5\xE6\xE7\xE8\xE9\xEA\xEB\xEC\xED\xEE\xEF'
+    '\xF0\xF1\xF2\xF3\xF4\xF5\xF6\xF7\xF8\xF9\xFA\xFB\xFC\xFD\xFE\xFF';
+
+// Characters of ISO-8859-16.
+const iso8859_16 = // From https://unicode.org/Public/MAPPINGS/ISO8859/8859-16.TXT
+    '\x00\x01\x02\x03\x04\x05\x06\x07\b\t\n\v\f\r\x0E\x0F'
+    '\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F'
+    ' !"#\$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_'
+    '`abcdefghijklmnopqrstuvwxyz{|}~\x7F'
+    '\x80\x81\x82\x83\x84\x85\x86\x87\x88\x89\x8A\x8B\x8C\x8D\x8E\x8F'
+    '\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9A\x9B\x9C\x9D\x9E\x9F'
+    '\xA0\u0104\u0105\u0141\u20AC\u201E\u0160\xA7'
+    '\u0161\xA9\u0218\xAB\u0179\xAD\u017A\u017B'
+    '\xB0\xB1\u010C\u0142\u017D\u201D\xB6\xB7'
+    '\u017E\u010D\u0219\xBB\u0152\u0153\u0178\u017C'
+    '\xC0\xC1\xC2\u0102\xC4\u0106\xC6\xC7\xC8\xC9\xCA\xCB\xCC\xCD\xCE\xCF'
+    '\u0110\u0143\xD2\xD3\xD4\u0150\xD6\u015A'
+    '\u0170\xD9\xDA\xDB\xDC\u0118\u021A\xDF\xE0\xE1\xE2\u0103\xE4\u0107\xE6\xE7'
+    '\xE8\xE9\xEA\xEB\xEC\xED\xEE\xEF\u0111\u0144\xF2\xF3\xF4\u0151\xF6\u015B'
+    '\u0171\xF9\xFA\xFB\xFC\u0119\u021B\xFF';
+// -- END GENERATED CONTENT --